refactor: splice dna_utils.py out of audio_utils.py
