+++ test/test_anchors.py
@@ -12,7 +12,7 @@ class TestAnchors(TestCase):
 
     # Set this to True to keep the temp dirs after running, for manual inspection
     # but please don't push a commit setting this to True!
-    keep_temp_dir_after_running = True
+    keep_temp_dir_after_running = False
     if not keep_temp_dir_after_running:
         tempdirobj = tempfile.TemporaryDirectory(prefix="tmpdir_test_anchors_", dir=".")
         tempdir = tempdirobj.name
